deprecate old trade route; redirec to multi-trade flow
